plugin architecture fleshed out and overhauled further.
Current limitations, only sort plugins allowed for and only one can be enabled at a time.  That makese sense for Sorts but maybe not for other types.

Currently the BagBreak, Valign and BagOrder options won't work as they aren't tied into the plugin yet.

Yes the 'BROKEN' sort plugin is a test plugin, and yes it'll be removed.